% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbDenseMatrix-class}
\alias{dbDenseMatrix-class}
\alias{dbDenseMatrix}
\title{S4 Class for \code{dbDenseMatrix}}
\value{
Objects of class \code{dbDenseMatrix} store all matrix entries explicitly
in DuckDB. They are typically returned by \code{\link[=dbMatrix]{dbMatrix()}} or \code{\link[=as.dbMatrix]{as.dbMatrix()}}
for dense inputs.
}
\description{
Representation of dense matrices using an on-disk database.
Inherits from \link{dbMatrix}.
}
