% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/writeMM.R
\name{dbLoad,DBIConnection-method}
\alias{dbLoad,DBIConnection-method}
\alias{writeMM,dbMatrix-method}
\title{\subsection{dbDisconnect}{

#' @title dbDisconnect
#' @rdname DBI
#' @export
setMethod('dbDisconnect', signature(x = 'dbMatrix'),
function(x, ...){
con <- get_con(x)
DBI::dbDisconnect(conn = con, shutdown = TRUE)
})
}}
\usage{
\S4method{dbLoad}{DBIConnection}(conn, name, class)

\S4method{writeMM}{dbMatrix}(obj, file, ...)
}
\arguments{
\item{conn}{DBIConnection object}

\item{name}{valid name value (character)}

\item{class}{character, class of the dbMatrix object (e.g. "dbDenseMatrix" or "dbSparseMatrix")}

\item{obj}{dbMatrix object}

\item{file}{path to file}

\item{...}{additional arguments}
}
\description{
\subsection{dbListTables}{

#' @title dbListTables
#' @rdname DBI
#' @export
setMethod('dbListTables', signature(x = 'dbMatrix'),
function(x, ...){
con <- get_con(x)
DBI::dbListTables(conn = con)
})
Create a dbMatrix object computed in a database
}
}
