% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{dbMatrix-options}
\alias{dbMatrix-options}
\alias{dbMatrix.dbdm_auto_compute}
\title{dbMatrix Package Global Options}
\description{
The following global options can be modified to control the
behavior of the \code{\link{dbMatrix}} package.
}
\details{
Use \code{options()} to set the below options.
}
\section{Options}{

\itemize{
\item \code{\link{dbMatrix.dbdm_auto_compute}}: logical. If \code{TRUE}, automatically
computes intermediate dense COO tables that are generated from
\code{\link{dbSparseMatrix}} to \code{\link{dbDenseMatrix}} conversion in the internal
function \code{\link{.to_db_dense}}.
If \code{FALSE} (default), this step is skipped and the dense COO table is lazy.
Setting this to \code{TRUE} is recommended to avoid large intermediate COO tables
that can trigger downstream bottlenecks or memory errors due to disk spilling
when using DuckDB as a backend.
\item \code{\link{dbMatrix.digits}}: integer. Number of digits to round to in the show
function of dbMatrix objects. Default is 7.
}
}

\keyword{internal}
