% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{dim,dbMatrix-method}
\alias{dim,dbMatrix-method}
\title{Dimensions of an Object}
\usage{
\S4method{dim}{dbMatrix}(x)
}
\arguments{
\item{x}{\code{\link{dbMatrix}} object}
}
\value{
An integer vector of length 2 giving the number of rows and columns
in \code{x}.
}
\description{
Retrieve the dimension of an object.
}
\concept{matrix_props}
