% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.check_tbl}
\alias{.check_tbl}
\title{Input validation for tbl arg}
\usage{
.check_tbl(tbl)
}
\arguments{
\item{tbl}{A table object (should be tbl_duckdb_connection)}
}
\description{
Input validation for tbl arg
}
\keyword{internal}
