% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.check_value}
\alias{.check_value}
\title{Input validation for data arg}
\usage{
.check_value(value)
}
\arguments{
\item{value}{A \code{Matrix}, \code{matrix}, or \code{tbl_duckdb_connection} object}
}
\value{
No return value. Called for input validation and throws an error if
\code{value} is invalid.
}
\description{
Input validation for data arg
}
\keyword{internal}
