% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{.eval_op_densify}
\alias{.eval_op_densify}
\title{Evaluate if a dbSparseMatrix should be densified}
\usage{
.eval_op_densify(generic_char, vec_matrix)
}
\arguments{
\item{generic_char}{A character string representing the operation to be performed.}

\item{vec_matrix}{A \code{dbMatrix} object with 1D row or col.}
}
\value{
\code{TRUE} if the operation should densify the sparse matrix before
evaluation, otherwise \code{FALSE}.
}
\description{
Evaluate if a dbSparseMatrix should be densified
}
\details{
Evaluates if a \code{dbSparseMatrix} should be
densified for \verb{[Arith]} operations and specific scalar values for
operations in the order of dbSparseMatrix, vector
}
\keyword{internal}
