% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{.join_dbm_vector}
\alias{.join_dbm_vector}
\title{Join a \code{dbSparseMatrix} with a \code{dbMatrix} object}
\usage{
.join_dbm_vector(dbm, dbVector, op, swap_arith_order = FALSE)
}
\arguments{
\item{dbm}{A \code{dbSparseMatrix} object.}

\item{dbVector}{A \code{dbMatrix} object with 1D row or col.}

\item{swap_arith_order}{order of the arguments for the operation. default: NULL}

\item{generic_char}{A character string representing the operation to be performed.}
}
\description{
Join a \code{dbSparseMatrix} with a \code{dbMatrix} object
}
\keyword{internal}
