% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_con}
\alias{get_con}
\title{get_con}
\usage{
get_con(dbMatrix)
}
\arguments{
\item{dbMatrix}{A database-backed object inheriting from \code{dbData}.}
}
\value{
A live DBI connection associated with the database-backed object.
}
\description{
get_con
}
\keyword{internal}
