% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{get_MM_dimnames}
\alias{get_MM_dimnames}
\title{get_MM_dimnames}
\usage{
get_MM_dimnames(
  mtx_file_path,
  mtx_rowname_file_path,
  mtx_rowname_col_idx = 1,
  mtx_colname_file_path,
  mtx_colname_col_idx = 1,
  ...
)
}
\arguments{
\item{mtx_file_path}{path to .mtx file to be read into database}

\item{mtx_rowname_file_path}{path to .mtx rowname file to be read into
database. by default, no header is assumed.}

\item{mtx_rowname_col_idx}{column index of row name file}

\item{mtx_colname_file_path}{path to .mtx colname file to be read into
database. by default, no header is assumed.}

\item{mtx_colname_col_idx}{column index of column name file}

\item{...}{additional params to pass to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
list of row and column name character vectors
}
\description{
Internal function to read row and column names of a .mtx file
}
\details{
Can be used to read row and column names from .mtx files. Note: these files
must not contain a header (colnames).

The mtx_rowname_col_idx and mtx_colname_col_idx can be used to specify the column
index of the row and column name files, respectively. By default, the first
column is used for both.

TODO: Support for reading in only rownames or colnames.
}
\keyword{internal}
