% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{map_ijx_dimnames}
\alias{map_ijx_dimnames}
\title{Map dimnames to i,j indices}
\usage{
map_ijx_dimnames(dbMatrix, colName_i, colName_j)
}
\arguments{
\item{dbMatrix}{dbMatrix object}

\item{colName_i}{name of column rownames to add to database}

\item{colName_j}{name of column colnames to add to database
default: 'FALSE'.'}
}
\value{
A lazy \code{tbl_dbi} with \code{i}, \code{j}, \code{x}, and the mapped row/column name
columns.
}
\description{
Map dimnames to i,j indices
}
\details{
Constructs a table in a database that contains the accompanying dimnames
for a dbMatrix. The resulting columns in the table:
\itemize{
\item i (row index)
\item colName_i (rownames),
\item j (col index)
\item j_names (colnames)
\item x (counts of i,j occcurences)
}
}
\keyword{internal}
