% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{rownames.dbMatrix}
\alias{rownames.dbMatrix}
\alias{rownames<-.dbMatrix}
\alias{colnames.dbMatrix}
\alias{colnames<-.dbMatrix}
\alias{dimnames,dbMatrix-method}
\alias{dimnames<-,dbMatrix,list-method}
\title{Retrieve and Set Row (Column) Dimension Names of dbMatrix Objects}
\usage{
rownames.dbMatrix(x, do.NULL = TRUE, prefix = "row")

\method{rownames}{dbMatrix}(x) <- value

colnames.dbMatrix(x, do.NULL = TRUE, prefix = "col")

\method{colnames}{dbMatrix}(x) <- value

\S4method{dimnames}{dbMatrix}(x)

\S4method{dimnames}{dbMatrix,list}(x) <- value
}
\arguments{
\item{x}{a matrix-like \R object, with at least two dimensions for
    \code{colnames}.}

\item{do.NULL}{Not used for this method. Included for compatibility with the
generic.}

\item{prefix}{Not used for this method. Included for compatibility with the
generic.}

\item{value}{a valid value for that component of
    \code{\link[base]{dimnames}(x)}.  For a matrix or array this is either
    \code{NULL} or a character vector of non-zero length equal to the
    appropriate dimension.}
}
\description{
Retrieve and Set Row (Column) Dimension Names of dbMatrix Objects
}
\concept{matrix_props}
