% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names,dbDenseMatrix-method}
\alias{names,dbDenseMatrix-method}
\title{The names of a dbMatrix Object}
\usage{
\S4method{names}{dbDenseMatrix}(x)
}
\arguments{
\item{x}{A dbMatrix object}
}
\value{
A character vector of the names of the 1D dbMatrix object (1D matrices only)
}
\description{
The names of a dbMatrix Object
}
\concept{matrix_props}
