% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{nrow.dbMatrix}
\alias{nrow.dbMatrix}
\alias{ncol.dbMatrix}
\title{The Number of Rows/Columns of a dbMatrix Object}
\usage{
nrow.dbMatrix(x)

ncol.dbMatrix(x)
}
\arguments{
\item{x}{\code{\link{dbMatrix}} object}
}
\value{
A length-one integer giving the number of rows or columns in \code{x}.
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns present in \code{x}.
}
\concept{matrix_props}
