% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{rowVars,dbDenseMatrix-method}
\alias{rowVars,dbDenseMatrix-method}
\alias{rowVars,dbSparseMatrix-method}
\alias{colVars,dbDenseMatrix-method}
\alias{colVars,dbSparseMatrix-method}
\title{Row (column) variances for \code{\link{dbMatrix}} objects}
\usage{
\S4method{rowVars}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = TRUE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{rowVars}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = TRUE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{colVars}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = TRUE,
  center = NULL,
  ...,
  useNames = TRUE
)

\S4method{colVars}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = TRUE,
  center = NULL,
  ...,
  useNames = TRUE
)
}
\arguments{
\item{x}{A \code{\link{dbMatrix}} object.}

\item{rows}{Always NULL for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{cols}{Always NULL for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{na.rm}{Always TRUE for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{center}{Always NULL for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{...}{Additional arguments (not used, but included for compatibility with the generic).}

\item{useNames}{Always TRUE for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}
}
\value{
A named numeric vector containing one sample variance per row or
column of \code{x}.
}
\description{
Calculates the variance for each row (column) of a matrix-like object.
}
\concept{summary}
