% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Summary,dbMatrix-method}
\alias{Summary,dbMatrix-method}
\title{Summary Methods for \code{\link{dbMatrix}} Objects}
\usage{
\S4method{Summary}{dbMatrix}(x, ..., na.rm = TRUE)
}
\arguments{
\item{x}{A dbMatrix object.}

\item{...}{Additional arguments (not used, but included for compatibility with the generic).}

\item{na.rm}{Logical. If TRUE, remove NA values before computation. Always set to TRUE for this implementation.}
}
\value{
The result of applying the respective summary function to the dbMatrix object.
The type of the return value depends on the specific function called.
}
\description{
Implements the \code{\link{S4groupGeneric}} group generic functions for dbMatrix objects.
}
\details{
This method provides implementations for the following \code{\link{S4groupGeneric}} functions:
\itemize{
\item \code{max()}: Maximum value
\item \code{min()}: Minimum value
\item \code{range()}: \emph{Not supported}
\item \code{prod()}: Product of all values
\item \code{sum()}: Sum of all values
\item \code{any()}: Returns TRUE if any value is TRUE
\item \code{all()}: Returns TRUE if all values are TRUE
}
}
\examples{
mat <- matrix(1, nrow = 3, ncol = 3)
dbmat <- as.dbMatrix(mat)
max(dbmat)
min(dbmat)
prod(dbmat)
sum(dbmat)
any(dbmat > 0)
all(dbmat > 0)

}
\concept{summary}
