% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{t,dbMatrix-method}
\alias{t,dbMatrix-method}
\title{Matrix Transpose}
\usage{
\S4method{t}{dbMatrix}(x)
}
\arguments{
\item{x}{\code{\link{dbMatrix}} object}
}
\value{
\code{\link{dbMatrix}} object
}
\description{
Given a \code{\link{dbMatrix}} \code{x}, \code{t} returns the transpose of \code{x}.
}
\concept{transform}
