% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_view,dbData-method}
\alias{to_view,dbData-method}
\title{Convert lazy table to named view}
\usage{
\S4method{to_view}{dbData}(x, name, temporary = TRUE, ...)
}
\arguments{
\item{x}{dbData object to convert to a view \code{(required)}}

\item{name}{view name to assign within database. auto-generated if none provided \code{(required)}}

\item{temporary}{if TRUE, the view will be removed after the session ends}
}
\description{
Convert lazy table to named view
}
\concept{dbData}
