% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{as.dbMatrix}
\alias{as.dbMatrix}
\title{Convert \code{\link{Matrix}} to \code{\link{dbMatrix}}}
\usage{
as.dbMatrix(x, con, name, ...)
}
\arguments{
\item{x}{\code{\link{matrix}}, \code{\link[Matrix:dgeMatrix-class]{Matrix::dgeMatrix}}, or \code{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}} \code{required}}

\item{con}{\code{\link{tbl_duckdb_connection}} \code{default:"memory"} Connection to
DuckDB database connection. If not provided, a temporary in-memory
DuckDB database is created. \code{':temp:'}  will create a DuckDB database
in the temporary directory. \code{':memory:'} will create a DuckDB database
in memory.}

\item{name}{\code{default:"memory"} table name in the database. If not
provided, a unique table name is generated.}

\item{...}{Additional arguments passed to \code{\link{dbMatrix}}}
}
\description{
Converts in-memory \code{\link{matrix}}, \code{\link[Matrix:dgeMatrix-class]{Matrix::dgeMatrix}}, or
\code{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}} into a \code{\link{dbMatrix}} object.
}
\details{
If no \code{con} is provided, a temporary in-memory database connection is created.
If no \code{name} is provided, a unique table name is generated.
}
\concept{dbMatrix}
