% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coerce-dbMatrix-dgCMatrix}
\alias{coerce-dbMatrix-dgCMatrix}
\alias{coerce-dbDenseMatrix-dgCMatrix}
\alias{coerce-dbSparseMatrix-dgCMatrix}
\title{Coerce dbMatrix to dgCMatrix}
\description{
Coercion methods to convert \code{dbMatrix} objects to in-memory \code{dgCMatrix} objects.
Respects \code{dbMatrix.max_mem_convert} option to prevent OOM errors.
}
