% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coerce-dbMatrix-matrix}
\alias{coerce-dbMatrix-matrix}
\alias{coerce-dbDenseMatrix-matrix}
\alias{coerce-dbSparseMatrix-matrix}
\title{Coerce dbMatrix to matrix}
\description{
Coercion methods to convert \code{dbMatrix} objects to in-memory \code{matrix} objects.
Respects \code{dbMatrix.max_mem_convert} option to prevent OOM errors.
}
