% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbDenseMatrix-class}
\alias{dbDenseMatrix-class}
\alias{dbDenseMatrix}
\title{S4 Class for \code{dbDenseMatrix}}
\description{
Representation of dense matrices using an on-disk database.
Inherits from \link{dbMatrix}.
}
