% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{dbMatrix-class}
\alias{dbMatrix-class}
\alias{dbMatrix}
\title{S4 virtual class for \code{dbMatrix}}
\usage{
dbMatrix(
  value,
  class = NULL,
  con = NULL,
  overwrite = FALSE,
  name = "dbMatrix",
  dims = NULL,
  dim_names = NULL,
  mtx_rowname_file_path,
  mtx_rowname_col_idx = 1,
  mtx_colname_file_path,
  mtx_colname_col_idx = 1,
  ...
)
}
\description{
Representation of sparse and dense matrices in a database. Each object
is used as a connection to a single table that exists within the database.
Inherits from \code{dbData}.
}
\section{Slots}{

\describe{
\item{\code{dim_names}}{row \link{1} and col \link{2} names}

\item{\code{dims}}{dimensions of the matrix}

\item{\code{init}}{logical. Whether the object is fully initialized}
}}

