% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R, R/utils.R, R/writeMM.R
\name{Arith,dbMatrix,ANY-method}
\alias{Arith,dbMatrix,ANY-method}
\alias{Arith,ANY,dbMatrix-method}
\alias{Arith,dbMatrix,dbMatrix-method}
\alias{Ops,dbMatrix,ANY-method}
\alias{Ops,ANY,dbMatrix-method}
\alias{Ops,dbMatrix,dbMatrix-method}
\alias{dbLoad,DBIConnection-method}
\alias{writeMM,dbMatrix-method}
\title{Arith dbMatrix, e2}
\usage{
\S4method{Arith}{dbMatrix,ANY}(e1, e2)

\S4method{Arith}{ANY,dbMatrix}(e1, e2)

\S4method{Arith}{dbMatrix,dbMatrix}(e1, e2)

\S4method{Ops}{dbMatrix,ANY}(e1, e2)

\S4method{Ops}{ANY,dbMatrix}(e1, e2)

\S4method{Ops}{dbMatrix,dbMatrix}(e1, e2)

\S4method{dbLoad}{DBIConnection}(conn, name, class)

\S4method{writeMM}{dbMatrix}(obj, file, ...)
}
\arguments{
\item{e1}{First operand.}

\item{e2}{Second operand.}

\item{conn}{DBIConnection object}

\item{name}{valid name value (character)}

\item{class}{character, class of the dbMatrix object (e.g. "dbDenseMatrix" or "dbSparseMatrix")}

\item{obj}{dbMatrix object}

\item{file}{path to file}

\item{...}{additional arguments}
}
\description{
See \code{methods::Arith} for more details.

See \code{methods::Arith} for more details.

See \code{methods::Arith} for more details.

See \code{methods::Ops} for more details.

See \code{methods::Ops} for more details.

See \code{methods::Ops} for more details.
}
