% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{.join_dbm_vect}
\alias{.join_dbm_vect}
\title{Join a \code{dbSparseMatrix} with a \code{dbMatrix} object}
\usage{
.join_dbm_vect(dbm, vec_matrix, op, swap_arith_order = FALSE)
}
\arguments{
\item{dbm}{A \code{dbSparseMatrix} object.}

\item{vec_matrix}{A \code{dbMatrix} object with 1D row or col.}

\item{op}{A character string representing the operation to be performed.}

\item{swap_arith_order}{order of the arguments for the operation. default: NULL}
}
\value{
A \code{\link{dbMatrix}} object containing the result of applying \code{op} between
\code{dbm} and \code{vec_matrix}.
}
\description{
Join a \code{dbSparseMatrix} with a \code{dbMatrix} object
}
\keyword{internal}
