% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix.R
\name{.to_db_dense}
\alias{.to_db_dense}
\title{Convert a dbSparseMatrix to dbDenseMatrix}
\usage{
.to_db_dense(x)
}
\arguments{
\item{x}{A \code{\link{dbSparseMatrix}} object}
}
\value{
A \code{\link{dbDenseMatrix}} object
}
\description{
Internal function to convert a \code{\link{dbSparseMatrix}} to
\code{\link{dbDenseMatrix}}.
}
\examples{
dbsm <- sim_dbSparseMatrix(10,10)
dbdm <- toDbDense(dbsm)
}
\keyword{internal}
