% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{length,dbMatrix-method}
\alias{length,dbMatrix-method}
\title{Length of a \code{\link{dbMatrix}} Object}
\usage{
\S4method{length}{dbMatrix}(x)
}
\arguments{
\item{x}{\code{\link{dbMatrix}} object}
}
\description{
Get or set the length of vectors (including lists) and factors,
and of any other R object for which a method has been defined.
}
\concept{matrix_props}
