% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mean,dbDenseMatrix-method}
\alias{mean,dbDenseMatrix-method}
\alias{mean,dbSparseMatrix-method}
\title{Arithmetic Mean for \code{\link{dbMatrix}} objects}
\usage{
\S4method{mean}{dbDenseMatrix}(x, ...)

\S4method{mean}{dbSparseMatrix}(x, ...)
}
\arguments{
\item{x}{\code{\link{dbMatrix}} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function for the (trimmed) arithmetic mean.
}
\concept{summary}
