% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{nrow,dbMatrix-method}
\alias{nrow,dbMatrix-method}
\alias{ncol,dbMatrix-method}
\title{The Number of Rows/Columns of a dbMatrix Object}
\usage{
\S4method{nrow}{dbMatrix}(x)

\S4method{ncol}{dbMatrix}(x)
}
\arguments{
\item{x}{\code{\link{dbMatrix}} object}
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns present in \code{x}.
}
\concept{matrix_props}
