% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{rowMeans,dbMatrix-method}
\alias{rowMeans,dbMatrix-method}
\alias{colMeans,dbMatrix-method}
\title{Row (column) means for \code{\link{dbMatrix}} objects}
\usage{
\S4method{rowMeans}{dbMatrix}(x, na.rm = FALSE, dims = 1, ...)

\S4method{colMeans}{dbMatrix}(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x}{An NxK matrix-like object, a numeric data frame, or an array-like
object of two or more dimensions.}

\item{na.rm}{Always TRUE for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{dims}{Always 1 for \code{\link{dbMatrix}} queries. Included for compatibility with
the generic.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A named numeric vector containing one mean per row or column of \code{x}.
}
\description{
Calculates the mean for each row (column) of a matrix-like object.
}
\concept{summary}
