% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{rowMeans,dbMatrix-method}
\alias{rowMeans,dbMatrix-method}
\alias{colMeans,dbMatrix-method}
\title{Row (column) means for \code{\link{dbMatrix}} objects}
\usage{
\S4method{rowMeans}{dbMatrix}(x, ..., memory = FALSE)

\S4method{colMeans}{dbMatrix}(x, ..., memory = FALSE)
}
\arguments{
\item{x}{An NxK matrix-like object, a numeric data frame, or an array-like
object of two or more dimensions.}

\item{...}{Additional arguments passed to specific methods.}

\item{memory}{logical. If FALSE (default), results returned as dbDenseMatrix.
This is recommended for large computations. Set to TRUE to return the results
as a vector.}

\item{na.rm}{Always TRUE for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{dims}{Always 1 for \code{\link{dbMatrix}} queries. Included for compatibility with
the generic.}
}
\description{
Calculates the mean for each row (column) of a matrix-like object.
}
\concept{summary}
