% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{colSds,dbDenseMatrix-method}
\alias{colSds,dbDenseMatrix-method}
\alias{colSds,dbSparseMatrix-method}
\alias{rowSds,dbDenseMatrix-method}
\alias{rowSds,dbSparseMatrix-method}
\title{Row (column) standard deviations for \code{\link{dbMatrix}} objects}
\usage{
\S4method{colSds}{dbDenseMatrix}(x, ..., memory = FALSE, useNames = TRUE)

\S4method{colSds}{dbSparseMatrix}(x, ..., memory = FALSE, useNames = TRUE)

\S4method{rowSds}{dbDenseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = TRUE,
  center = NULL,
  ...,
  memory = FALSE,
  useNames = TRUE
)

\S4method{rowSds}{dbSparseMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = TRUE,
  center = NULL,
  ...,
  memory = FALSE,
  useNames = TRUE
)
}
\arguments{
\item{x}{A \code{\link{dbMatrix}} object.}

\item{...}{Additional arguments (not used, but included for compatibility with the generic).}

\item{memory}{logical. If FALSE (default), results returned as dbDenseMatrix.}

\item{useNames}{Always TRUE for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{rows}{Always NULL for \code{\link{dbMatrix}} queries. TODO}

\item{cols}{Always NULL for \code{\link{dbMatrix}} queries. TODO}

\item{na.rm}{Always TRUE for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}

\item{center}{Always NULL for \code{\link{dbMatrix}} queries. Included for compatibility
with the generic.}
}
\description{
Calculates the standard deviation for each row (column) of a matrix-like
object.
}
\concept{summary}
